backend/requirements.txt
numpy 
scipy 
scapy
psutil 
pandas 
numpy 
eventlet 
python-socketio 
scikit-learn
requests